refactor mysql-lib to keep unified module style.
